+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 - [#219](https://github.com/smile-io/ember-polaris/pull/219) [ENHANCEMENT] Update `polaris-date-picker` components to match Polaris v2.11.0
 - [#221](https://github.com/smile-io/ember-polaris/pull/221) [ENHANCEMENT] Add `disabled` attribute to `polaris-choice` in `polaris-radio-button`
 - [#226](https://github.com/smile-io/ember-polaris/pull/226) [FEATURE] Add `polaris-sticky` component
+- [#227](https://github.com/smile-io/ember-polaris/pull/227) [ENHANCEMENT] Update `polaris-text-field` to match Polaris v2.11.0
+- [#230](https://github.com/smile-io/ember-polaris/pull/230) [ENHANCEMENT] Add `fullWidth` and `connectedTop` to `polaris-button-group`
 
 ### v1.7.8 (October 10, 2018)
 - [199](https://github.com/smile-io/ember-polaris/pull/199) [ENHANCEMENT] Add support for disabled and loading states to `polaris-setting-toggle`.